rbe node: split up tests
